Remove old unused you-are-here-icon code
Basically, the radius circle is enough, especially with the drop-
marker attached to its center by default. "You are here" is really
just "you are somewhere around here", and the user knows exactly
where in the circle they are (that's what the drop marker is for).

I may include a nav icon in the main map, but not the dropmap.